Migrate structured types tests to new file
