+++ src/test/java/net/snowflake/client/jdbc/ArrowResultSetStructuredTypesLatestIt.java
@@ -1,0 +1,7 @@
+package net.snowflake.client.jdbc;
+
+public class ArrowResultSetStructuredTypesLatestIt extends BaseResultSetStructuredTypesLatestIT {
+  public ArrowResultSetStructuredTypesLatestIt() {
+    super("ARROW");
+  }
+}
