+++ src/test/java/net/snowflake/client/jdbc/JsonResultSetStructuredTypesLatestIt.java
@@ -1,0 +1,7 @@
+package net.snowflake.client.jdbc;
+
+public class JsonResultSetStructuredTypesLatestIt extends BaseResultSetStructuredTypesLatestIT {
+  public JsonResultSetStructuredTypesLatestIt() {
+    super("JSON");
+  }
+}
